+++ src/esm/drawgl.js
@@ -556,7 +556,7 @@ export class drawutilsgl {
         // NOT YET IMPLEMENTED
     }
     /**
-     * Draw a text label at the given relative position.
+     * Draw a text at the given relative position.
      *
      * @method text
      * @param {string} text - The text to draw.
@@ -585,14 +585,12 @@ export class drawutilsgl {
      * @param {number} x - The x-position to draw the text at.
      * @param {number} y - The y-position to draw the text at.
      * @param {number=} rotation - The (aoptional) rotation in radians.
+     * @param {string="black"} color - The color to use (default is black).
      * @return {void}
      * @instance
      * @memberof drawutils
      */
-    // +---------------------------------------------------------------------------------
-    // | Draw a non-scaling text label at the given position.
-    // +-------------------------------
-    label(text, x, y, rotation) {
+    label(text, x, y, rotation, color) {
         // NOT YET IMPLEMENTED
     }
     /**